move environment variable expansion to a separate file
