+++ apollo-router/src/configuration/tests.rs
@@ -1,4 +1,3 @@
-
 use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
